% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_brasilia.R
\name{data}
\alias{data}
\title{Monthly Wind Speed Data in Brasilia City}
\format{
A data frame with columns:
  - \code{Month}: The month of the observation.
  - \code{WindSpeed}: The recorded wind speed in meters per second.
}
\usage{
data()
}
\value{
Um dataframe contendo os dados
}
\description{
This dataset contains monthly wind speed data recorded in Brasilia city.
}
