% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmax.R
\name{dmax}
\alias{dmax}
\title{Probability Density Function of the Parametrized Maxwell Distribution}
\usage{
dmax(y, mu)
}
\arguments{
\item{y}{A numeric value or vector at which the density is to be evaluated. Must be non-negative.}

\item{mu}{A positive numeric value representing the scale parameter \eqn{\mu}.}
}
\value{
A numeric vector with the values of the density function.
}
\description{
Computes the probability density function (PDF) of a random variable \eqn{Y}
that follows a Maxwell distribution with a scale parameter \eqn{\mu > 0}.
The density function is given by:
}
\details{
\deqn{f(y;\mu) = \frac{32y^2}{\pi^2\mu^3} \exp\left(-\frac{4y^2}{\pi\mu^2}\right)}

This distribution is defined for \eqn{y \ge 0}.
}
\examples{
dmax(y = 1, mu = 2)

# Vectorized example
dmax(y = c(0.5, 1, 2), mu = 2)

# Plotting the PDF for mu = 3
curve(dmax(x, 3), from = 0, to = 10,
      xlab = "y", ylab = "Density", n = 1000)

}
