% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitmarma.R
\name{mxarma.fit}
\alias{mxarma.fit}
\title{MxARMA fitting without regressors}
\usage{
mxarma.fit(y, ar = NA, ma = NA, resid = 1, h1 = 0)
}
\arguments{
\item{y}{The vector of variables to fit}

\item{ar}{are the indices for the auto-regression.}

\item{ma}{are the indices for the moving-averages}

\item{resid}{Residual type for the model (1 for quantile residuals, 2 for standardized residuals, 3 for alternative residuals).}

\item{h1}{Number of predicted observations.}
}
\description{
Used for fitting the Maxwell auto-regressive moving averages model without regressors. The model is given by
}
\details{
\deqn{\eta_t = \alpha + \sum_{j\in ar}\phi_j \log(y_{t-j}) +\sum_{j\in ma}\theta_jr_{t-j}}
Where \itemize{
\item{\eqn{y} are the variables}
\item{\eqn{\alpha} is the intercept}
\item{\eqn{ar} are the indices for the auto-regression}
\item{\eqn{ma} are the indices for the moving-averages}
\item{\eqn{\phi} are the auto-regression coefficients}
\item{\eqn{\theta} are the moving-averages coefficients}
\item{\eqn{r} are the errors}}
}
\examples{
y <- MxARMA::mxarma.sim(100, alpha=0.6, phi= c(0.6,0.2), theta = -0.4)
mxarma.fit(y, ar=c(1,2), ma=c(1), resid = 1, h1 = 12)


}
