% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmax.R
\name{qmax}
\alias{qmax}
\title{Quantile Function of the Parametrized Maxwell Distribution}
\usage{
qmax(u, mu)
}
\arguments{
\item{u}{A numeric value or vector of probabilities in \eqn{[0, 1]}.}

\item{mu}{A positive numeric value representing the scale parameter \eqn{\mu}.}
}
\value{
A numeric vector of quantiles corresponding to the given probabilities.
}
\description{
Computes the quantile function of the Maxwell distribution with a scale parameter \eqn{\mu > 0}.
It returns the value \eqn{y} such that \eqn{P(Y \le y) = u}.
}
\examples{
qmax(u = 0.5, mu = 2)
qmax(u = c(0.25, 0.5, 0.75), mu = 2)

}
