% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmax.R
\name{rmax}
\alias{rmax}
\title{Random Generation from the Parametrized Maxwell Distribution}
\usage{
rmax(n = 1, mu)
}
\arguments{
\item{n}{Number of observations to generate. Must be a positive integer.}

\item{mu}{A positive numeric value representing the scale parameter \eqn{\mu}.}
}
\value{
A numeric vector of random values from the Maxwell distribution.
}
\description{
Generates \code{n} random values from the Maxwell distribution with a given scale parameter \eqn{\mu > 0},
using the quantile function.
}
\examples{
rmax(n = 5, mu = 2)

}
