% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_brasilia.R
\docType{data}
\name{wind_speed}
\alias{wind_speed}
\title{Monthly Wind Speed Data in Brasília City}
\format{
A data frame with the following columns:
\describe{
  \item{Date}{The date of the observation.}
  \item{Wind_speed_mean}{The mean wind speed (m/s) recorded on that date.}
  \item{Wind_speed_max}{The maximum wind speed (m/s) recorded on that date.}
}
}
\source{
Instituto Nacional de Meteorologia (INMET) — Brasília Weather Station.
}
\usage{
wind_speed
}
\description{
This dataset contains monthly wind speed data recorded in Brasília city.
}
\examples{
data("wind_speed")
head(wind_speed)
}
\keyword{datasets}
