% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_brasilia.R
\docType{data}
\name{wind_speed_brasilia}
\alias{wind_speed_brasilia}
\title{Monthly Wind Speed Data in Brasilia City}
\format{
A data frame with columns:
  - \code{Month}: The month of the observation.
  - \code{WindSpeed}: The recorded wind speed in meters per second.
}
\usage{
wind_speed_brasilia
}
\value{
Um dataframe contendo os dados
}
\description{
This dataset contains monthly wind speed data recorded in Brasilia city.
}
\examples{
data("wind_speed_brasilia")
head(wind_speed_brasilia)

# Se você distribuiu o .txt junto (inst/extdata), dá para ler assim:
# fp <- system.file("extdata", "wind_speed_brasilia.txt", package = "MxARMA")
# read.delim(fp, sep = ";")
}
\keyword{datasets}
