% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simumarma.R
\name{mxarma.sim}
\alias{mxarma.sim}
\title{Simulation of the Maxwell Distribution}
\usage{
mxarma.sim(n, alpha = 0, phi = NULL, theta = NULL)
}
\arguments{
\item{n}{Number of random values to generate.}

\item{alpha}{A value for the intercept parameter.}

\item{phi}{A value or a vector of values for the regressor parameter phi.}

\item{theta}{A value or a vector of values for the moving average parameter theta.}
}
\description{
This function generates random values from the MxARMA model.
}
\details{
The function simulates values from the MxARMA model, which includes
 the possibility of generating values for an autoregressive (AR), moving average (MA),
  or combined autoregressive moving average (ARMA) model.
}
\examples{
 y <- mxarma.sim(n = 1000, alpha = 0.6, phi = c(0.6, 0.1), theta = 0.3)
 y

}
