% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxarmafit_reg.R
\name{mxarmareg.fit}
\alias{mxarmareg.fit}
\title{MxARMA fitting with regressors}
\usage{
mxarmareg.fit(y, cvar, ar = NA, ma = NA, resid = 1, h1 = 0, X_hat = 0)
}
\arguments{
\item{y}{The vector of variables to fit}

\item{cvar}{A matrix of covariates or regressors used in the model. It should have the appropriate dimensions for the computations.}

\item{ar}{Specified in the description.}

\item{ma}{Specified in the description.}

\item{resid}{Residual type for the model (1 for quantile residuals, 2 for standardized residuals, 3 for alternative residuals).}

\item{h1}{Number of predicted observations.}

\item{X_hat}{covariate prediction vector}
}
\description{
Used for fitting the Maxwell auto-regressive moving averages model with regressors. The model is given by
\deqn{\eta_t = \beta_{0} + \boldsymbol{x}_t^\top\boldsymbol{\beta}+
\sum_{j\in ar}\phi_j [\log(y_{t-j})-\boldsymbol{x}_{t-j}^\top\boldsymbol{\beta}]
+\sum_{j\in ma}\theta_jr_{t-j}}
}
\details{
#' Where \itemize{
\item{\eqn{y} are the variables}
\item{\eqn{\alpha} is the intercept}
\item{\eqn{\boldsymbol{x}} are the covariables}
\item{\eqn{\boldsymbol{\beta}} are the regression coefficients}
\item{\eqn{ar} are the indices for the auto-regression}
\item{\eqn{ma} are the indices for the moving-averages}
\item{\eqn{\phi} are the auto-regression coefficients}
\item{\eqn{\theta} are the moving-averages coefficients}
\item{\eqn{r} are the errors}}

The function fitted values from the MxARMA model, including the possibility of fitting values for an autoregressive with exogenous variables (ARX), moving average (MAX), or combined autoregressive moving average with exogenous variables (ARMAX) model.
}
\examples{

 X = stats::runif(1004, 0, 1)
 y <- mxarmareg.sim(n = 1000, alpha = 0.6, phi = c(0.6, 0.1), theta = 0.3, beta = 0.7, X = X)
 X <- X[5:1004]
 xhat <- matrix(runif(12, 0, 1), ncol = 1)
 mxarmareg.fit(y = y, cvar = X, ar = c(1,2), ma = c(1), resid = 1, h1 = 12, X_hat = xhat)


}
