% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxarmasim_reg.R
\name{mxarmareg.sim}
\alias{mxarmareg.sim}
\title{Simulation of the Maxwell Distribution with regressors}
\usage{
mxarmareg.sim(n, phi = NULL, theta = NULL, alpha = 0, beta = NULL, X = NULL)
}
\arguments{
\item{n}{Number of random values to generate.}

\item{phi}{A value or a vector of values for the autoregressive parameter phi.}

\item{theta}{A value or a vector of values for the moving average parameter theta.}

\item{alpha}{A value for the intercept parameter.}

\item{beta}{A value or a vector of values for the regressor coefficient.}

\item{X}{A matrix of covariates or regressors used in the model. It should have the appropriate dimensions for the computations.}
}
\description{
This function generates random values from the MxARMA model with regressors.
}
\details{
The function simulates values from the MxARMA model, including the possibility of generating values for an autoregressive with exogenous variables (ARX), moving average (MAX), or combined autoregressive moving average with exogenous variables (ARMAX) model.
}
\examples{

 X = stats::runif(1004, 0, 1)
 y <- mxarmareg.sim(n = 1000, alpha = 0.6, phi = c(0.6, 0.1), theta = 0.3, beta = 0.7, X = X)
 y

}
